requirements.txt
streamlit
black
google-generativeai